requirements.txt
Flask==3.0.3
flask-pymongo==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Flask-SocketIO==5.3.0
dnspython==2.6.0 
dotenv